+++ zerver/tests/test_e2ee_push_notifications.py
@@ -576,3 +576,65 @@ class RemovePushNotificationTest(E2EEPushNotificationTestCase):
                 f"Sent E2EE mobile push notifications for user {hamlet.id}: 1 via FCM, 1 via APNs",
                 zerver_logger.output[1],
             )
+
+
+class RequireE2EEPushNotificationsSettingTest(E2EEPushNotificationTestCase):
+    def test_content_redacted(self) -> None:
+        hamlet = self.example_user("hamlet")
+        aaron = self.example_user("aaron")
+        realm = hamlet.realm
+
+        message_id = self.send_personal_message(
+            from_user=aaron, to_user=hamlet, skip_capture_on_commit_callbacks=True
+        )
+        missed_message = {
+            "message_id": message_id,
+            "trigger": NotificationTriggers.DIRECT_MESSAGE,
+        }
+
+        realm.require_e2ee_push_notifications = True
+        realm.save(update_fields=["require_e2ee_push_notifications"])
+
+        # Verify that the content is redacted in payloads supplied to
+        # 'send_notifications_to_bouncer' - payload supplied to bouncer.
+        with (
+            activate_push_notification_service(),
+            mock.patch("zerver.lib.push_notifications.send_notifications_to_bouncer") as m,
+        ):
+            handle_push_notification(hamlet.id, missed_message)
+
+            m.assert_called_once()
+            self.assertEqual(m.call_args.args[1]["alert"]["body"], "New message")
+            self.assertEqual(m.call_args.args[2]["content"], "New message")
+
+        message_id = self.send_personal_message(
+            from_user=aaron, to_user=hamlet, skip_capture_on_commit_callbacks=True
+        )
+        missed_message = {
+            "message_id": message_id,
+            "trigger": NotificationTriggers.DIRECT_MESSAGE,
+        }
+
+        # Verify that the content is redacted in payloads supplied to
+        # to functions for sending it through APNs and FCM directly.
+        with (
+            mock.patch("zerver.lib.push_notifications.has_apns_credentials", return_value=True),
+            mock.patch("zerver.lib.push_notifications.has_fcm_credentials", return_value=True),
+            mock.patch(
+                "zerver.lib.push_notifications.send_notifications_to_bouncer"
+            ) as send_bouncer,
+            mock.patch(
+                "zerver.lib.push_notifications.send_apple_push_notification", return_value=0
+            ) as send_apple,
+            mock.patch(
+                "zerver.lib.push_notifications.send_android_push_notification", return_value=0
+            ) as send_android,
+        ):
+            handle_push_notification(hamlet.id, missed_message)
+
+            send_bouncer.assert_not_called()
+            send_apple.assert_called_once()
+            send_android.assert_called_once()
+
+            self.assertEqual(send_apple.call_args.args[2]["alert"]["body"], "New message")
+            self.assertEqual(send_android.call_args.args[2]["content"], "New message")